Do not wait for Basic.CancelOk response on blocking receive.
Pika 0.10 allows to pass empty CancelOk callback to the basic_cancel
method.  In that case it silently calls consumer tear-down logic.
Recently we upgrade to pika 0.11 which requires this call back to be
passed.  New library version handle consumer tear-down in other
places.  So it's safe to ignore CancelOk responses.